git-svn-id: file:///var/www/svn/mxtce-sw@7 52860be4-34ab-47e5-b986-9ec2afffd7f9

+++ src/common/event.hh
@@ -204,7 +204,7 @@ public:
 class EventMaster
 {
 private:
-    static map<u_int32_t, EventMaster*> threadMasters;
+    static map<unsigned long, EventMaster*> threadMasters;
     fd_set readfd;
     fd_set writefd;
     fd_set exceptfd;
